refactor(negotiation): negotiationHasMadeContact moves to the protocol module
The predicate is not a screening helper. `negotiation.graph.finalize.ts` and
`negotiation.graph.turn.ts` both call it for live turn logic — the IND-564
opening-`withdraw` guard and the `screened_out` label — and neither has
anything to do with the outreach gate. It lived in `negotiation.screen.contracts.ts`
by accident of history.

Pure move, no behaviour change: the function body is unchanged, its doc drops
the screen-gate framing for the two rules that actually depend on it, and its
unit coverage moves from `negotiation.screen.spec.ts` to
`negotiation.protocol.spec.ts` alongside it.
