+++ packages/protocol/src/negotiations/negotiation.graph.ts
@@ -15,8 +15,7 @@ import { NegotiationGraphState } from "./negotiation.state.js";
 import { IndexNegotiator } from "./negotiation.agent.js";
 import { NegotiationStallGapAuthor } from "./negotiation.stall-gap.js";
 import { blocksNegotiationBeforeFirstTurn, NegotiationScreener } from "./negotiation.screen.js";
-import { negotiationHasMadeContact } from "./negotiation.screen.contracts.js";
-import { isTerminalAction } from "./negotiation.protocol.js";
+import { isTerminalAction, negotiationHasMadeContact } from "./negotiation.protocol.js";
 import { isNegotiationTurnCapReached } from "./negotiation.turn-cap.js";
 import type { QuestionerEnqueueFn } from "../questions/question.module.js";
 import type { ReflectEnqueueFn } from "./negotiation.reflect.js";
