+++ packages/protocol/src/negotiations/negotiation.screen.contracts.ts
@@ -77,40 +77,3 @@ export function blocksNegotiationBeforeFirstTurn(
     turnCount === 0
   );
 }
-
-/**
- * Whether a negotiation has already made contact — the fact the outreach gate
- * exists to decide, and therefore the fact that makes re-deciding it a
- * category error.
- *
- * The screen asks one question: "should we make first contact?". Once an agent
- * message for this negotiation is on the counterparty's thread, that question
- * has been answered by the world, and a `pass` here would end a negotiation the
- * counterparty was never given a chance to answer — while the outreach it
- * denies sits visibly above the notice saying no contact was ever made. This
- * was observed live: an error-stalled negotiation recovered through
- * `negotiation-run-existing` re-screened, passed, and flipped to `rejected`
- * with `screened_out` two hours after its outreach landed.
- *
- * Contact is a PERSISTED TURN, not a screen decision: a screen record is task
- * metadata, so a continuation whose only prior activity is a screen has still
- * sent nothing and is still genuinely pre-contact.
- *
- * `ask_user` is the one persisted turn that is not contact. It parks the
- * negotiation to ask the client's OWN principal a question before deciding to
- * reach out; nothing was ever addressed to the counterparty. This is the same
- * reading `isPreContactConsultResume` already applies to an all-`ask_user`
- * history, and keeping the two in step is what lets a pre-contact consult
- * resume still resolve as the opening decision it is.
- *
- * Scope matters as much as the predicate: callers pass THIS negotiation's turns
- * (`readNegotiationMessages`), never the pair's whole DM. A new match reusing an
- * established `dm_pair` conversation has made no contact of its own and must
- * still be screened (IND-563) — the pair's earlier dialogue reaches the gate as
- * labelled context instead.
- */
-export function negotiationHasMadeContact(
-  turns: readonly { action: string }[],
-): boolean {
-  return turns.some((turn) => turn.action !== "ask_user");
-}
